feat: 약속 정보 firebase db 에 저장 #19
