+++ app/src/main/java/com/boosters/promise/DataSourceModule.kt
@@ -2,6 +2,9 @@ package com.boosters.promise
 
 import android.content.Context
 import androidx.room.Room
+import com.google.firebase.database.FirebaseDatabase
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
 import com.google.gson.Gson
 import dagger.Module
 import dagger.Provides
@@ -24,13 +27,19 @@ object DataSourceModule {
 
     @Provides
     @Singleton
-    fun provideSearchItemDatabase(
+    fun providePromiseLocalDatabase(
         @ApplicationContext appContext: Context,
         gson: Gson
     ): PromiseDatabase {
         return Room.databaseBuilder(appContext, PromiseDatabase::class.java, PROMISE_DATABASE)
             .addTypeConverter(UserTypeConverter(gson))
             .build()
+    }
+
+    @Provides
+    @Singleton
+    fun providePromiseRemoteDatabase(): FirebaseDatabase {
+        return Firebase.database("https://promise-c3e7d-default-rtdb.firebaseio.com/")
     }
 
 }
