+++ app/src/main/java/com/boosters/promise/PromiseLocalDataSourceImpl.kt
@@ -1,10 +1,19 @@
 package com.boosters.promise
 
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class PromiseLocalDataSourceImpl @Inject constructor(
     private val database: PromiseDatabase
 ) : PromiseLocalDataSource {
+
+    init {
+        CoroutineScope(Dispatchers.IO).launch {
+            database.promiseItemDao().deleteAll()
+        }
+    }
 
     override suspend fun addPromise(promise: Promise) {
         database.promiseItemDao().insert(promise)
