+++ app/src/main/java/com/boosters/promise/PromiseRepositoryImpl.kt
@@ -9,6 +9,7 @@ class PromiseRepositoryImpl @Inject constructor(
 
     override suspend fun addPromise(promise: Promise) {
         promiseLocalDataSource.addPromise(promise)
+        promiseRemoteDataSource.addPromise(promise)
     }
 
     override suspend fun removePromise(promise: Promise) {
